Improved output

+++ libs/evm/include/evm/data.hpp
@@ -31,7 +31,7 @@ namespace evm {
     struct EXPORT InputData {
         vector<Mat> _originals;
         vector<Roi> _rois;
-        int _fps;
+        double _fps;
         promise<OutputData> _promise;
     };
